Rebased. Fixed conflicts. Modified convert-key to produce the BLS pop-prove signature.
