+++ z2/src/validators.rs
@@ -169,17 +169,6 @@ fn hex_string_to_u8_20(hex_str: &str) -> Result<[u8; 20], &'static str> {
     Ok(array)
 }
 
-pub fn pop_prove(secret_key: &str) -> Result<()> {
-    let sk = hex::decode(secret_key)?;
-    if sk.len() != blsful::SECRET_KEY_BYTES {
-        return Err(anyhow!("Invalid secret key {}", secret_key));
-    }
-    let sk = blsful::SecretKey::<Bls12381G2Impl>::from_hash(sk);
-    let sk = sk.proof_of_possession()?;
-    println!("{}", hex::encode(sk.0.to_compressed()));
-    Ok(())
-}
-
 pub async fn get_chain_spec_config(chain_name: &str) -> Result<Value> {
     let contents = get_toml_contents(chain_name)?;
     let config: Value =
