+++ zilliqa/src/precompiles/erc20.rs
@@ -1,4 +1,4 @@
-use ethabi::{encode, ParamType, Token};
+use ethabi::{decode, encode, short_signature, ParamType, Token};
 use revm::{
     precompile::PrecompileError,
     primitives::{
@@ -7,7 +7,6 @@ use revm::{
     },
     ContextStatefulPrecompile, InnerEvmContext,
 };
-use sha3::{Digest, Keccak256};
 
 use crate::state::State;
 
@@ -19,7 +18,7 @@ impl ERC20Precompile {
         _gas_price: u64,
         context: &mut InnerEvmContext<&State>,
     ) -> PrecompileResult {
-        let Ok(decoded) = ethabi::decode(&[ParamType::Address], input) else {
+        let Ok(decoded) = decode(&[ParamType::Address], input) else {
             return Err(
                 PrecompileError::Other("Unable to decode provided account address".into()).into(),
             );
@@ -54,13 +53,6 @@ impl ERC20Precompile {
     }
 }
 
-fn make_selector(signature: &str) -> [u8; 4] {
-    let signature = Keccak256::digest(signature).to_vec();
-    let slice = signature.as_slice();
-    let res: [u8; 4] = slice[..4].try_into().unwrap();
-    res
-}
-
 impl ContextStatefulPrecompile<&State> for ERC20Precompile {
     fn call(
         &self,
@@ -75,8 +67,10 @@ impl ContextStatefulPrecompile<&State> for ERC20Precompile {
             .into());
         }
 
-        let dispatch_table: [([u8; 4], _); 1] =
-            [(make_selector("balanceOf(address)"), Self::get_balance)];
+        let dispatch_table: [([u8; 4], _); 1] = [(
+            short_signature("balanceOf", &[ParamType::Address]),
+            Self::get_balance,
+        )];
 
         let Some(handler) = dispatch_table
             .iter()
