+++ zilliqa/src/precompiles/mod.rs
@@ -1,10 +1,12 @@
 mod erc20;
+mod pop_verify;
 mod scilla_read;
 
 use std::sync::Arc;
 
 use alloy::primitives::Address;
 use erc20::ERC20Precompile;
+use pop_verify::PopVerify;
 use revm::ContextPrecompile;
 use scilla_read::ScillaRead;
 
@@ -15,6 +17,10 @@ pub fn get_custom_precompiles<'a>() -> Vec<(Address, ContextPrecompile<&'a State
         (
             Address::from(*b"\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0ZIL"),
             ContextPrecompile::ContextStateful(Arc::new(ERC20Precompile)),
+        ),
+        (
+            Address::from(*b"\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0PPV"),
+            ContextPrecompile::ContextStateful(Arc::new(PopVerify)),
         ),
         (
             Address::from(*b"\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0ZIL\x92"),
